+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Hero Detail feature presentation.
 - Hero Detail remote image loading.
 - Added mock generation script.
+- Routing target.
 
 ### Changed